Move ticket repo
