+++ src/main/java/ua/edu/ukma/cinemax/controller/CashierContoller.java
@@ -3,7 +3,7 @@ package ua.edu.ukma.cinemax.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ua.edu.ukma.cinemax.stat.repository.TicketRepository;
+import ua.edu.ukma.cinemax.repository.TicketRepository;
 
 @RestController
 public class CashierContoller {
