+++ src/main/java/ua/edu/ukma/cinemax/repository/Tickets.java
@@ -1,4 +1,4 @@
-package ua.edu.ukma.cinemax.stat.repository;
+package ua.edu.ukma.cinemax.repository;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
