+++ src/main/java/ua/edu/ukma/cinemax/stat/service/impl/TicketStatServiceImpl.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import ua.edu.ukma.cinemax.stat.repository.TicketRepository;
+import ua.edu.ukma.cinemax.repository.TicketRepository;
 import ua.edu.ukma.cinemax.stat.service.StatisticsService;
 
 import java.util.Date;
